backend/requirements.txt
# For Embedding Models
sentence-transformers==2.2.2
torch>=1.13.0
transformers>=4.36.0
numpy>=1.21.0

# For Text Chunking
langchain==0.1.15

# For Document Parsing
PyMuPDF==1.23.21      # (fitz)
python-docx==1.1.0

# For Text Processing
nltk==3.8.1

# MongoDB Driver (optional, for later DB interactions)
pymongo==4.5.0

python-dotenv==1.0.1
